// src/modules/session-vault/store/types.ts
export interface SessionKey {
  userId: string;
  family: string;
  sessionId: string;
}